Fix repeated creation of axis in loop
The previous code in this example creates x- and y-axis repeatedly. Once per
row in the CSV file. That results in thousands of SVG elements for the axes,
ticks, etc.

As given the previous code does not plot any nodes in the scatter plot for me.

This commit moves creation of axes out of the loop. Now, the resulting SVG
visualization looks as the one given on this HTML page.